api/requirements.txt
flask
prisma
pyrogram
tgcrypto
python-dotenv
requests
colorlog